@@ -3,50 +3,65 @@
   "description": "An implementation of the SQL Server Resolution protocol.",
   "version": "0.0.0-development",
   "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "scripts": {
-    "pretest": "eslint . && flow && pegjs -o src/parser.js src/parser.pegjs",
+    "pretest": "eslint . --ext js,ts && tsc && pegjs -o src/parser.js src/parser.pegjs",
     "test": "mocha",
-    "prepublish": "pegjs -o src/parser.js src/parser.pegjs && babel src --out-dir lib && flow-copy-source src lib",
-    "semantic-release": "semantic-release pre && npm publish && semantic-release post"
+    "build": "pegjs -o src/parser.js src/parser.pegjs",
+    "prepublish": "pegjs -o src/parser.js src/parser.pegjs && babel src --out-dir lib --extensions .js,.ts && tsc --noEmit false --emitDeclarationOnly true && cp src/parser.d.ts lib/parser.d.ts"
   },
   "author": "Arthur Schreiber (schreiber.arthur@gmail.com)",
   "license": "ISC",
   "devDependencies": {
-    "babel-cli": "^6.24.0",
-    "babel-eslint": "^7.2.1",
-    "babel-plugin-syntax-flow": "^6.18.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.2.2",
-    "babel-preset-flow": "^6.23.0",
-    "babel-register": "^6.24.0",
-    "chai": "^3.5.0",
-    "eslint": "^3.18.0",
-    "flow-bin": "^0.42.0",
-    "flow-copy-source": "^1.1.0",
-    "flow-typed": "^2.0.0",
-    "mocha": "^3.2.0",
+    "@babel/cli": "^7.10.1",
+    "@babel/core": "^7.10.2",
+    "@babel/node": "^7.10.1",
+    "@babel/plugin-transform-runtime": "^7.10.1",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-typescript": "^7.10.1",
+    "@babel/register": "^7.10.1",
+    "@types/chai": "^4.2.11",
+    "@types/chai-as-promised": "^7.1.2",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^14.0.11",
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "eslint": "^5.16.0",
+    "mocha": "^6.2.3",
     "pegjs": "^0.10.0",
     "semantic-release": "^6.3.2",
     "semantic-release-cli": "^3.0.3",
     "typescript": "^3.9.5"
   },
   "dependencies": {
-    "babel-runtime": "^6.23.0"
+    "@babel/runtime": "^7.0.0",
+    "abort-controller": "^3.0.0"
   },
   "babel": {
+    "ignore": [
+      "./src/*.d.ts"
+    ],
     "presets": [
       [
-        "env",
+        "@babel/preset-env",
         {
           "targets": {
-            "node": 4
-          }
+            "node": 6
+          },
+          "shippedProposals": true
         }
       ],
-      "flow"
+      [
+        "@babel/preset-typescript",
+        {
+          "allowDeclareFields": true
+        }
+      ]
     ],
     "plugins": [
-      "transform-runtime"
+      "@babel/plugin-transform-runtime"
     ]
   },
   "repository": {
